moved calorieTracker to blueprint & partially connected tracker ui to db
